added explicit utility class
